Add files via upload
the program analyzes a huge amount of data (voltages and phase angles from the entire 300 km long section of the line during an accident on the line) a parser was written in the Python programming language, which parsed the data and built the necessary graphs and waveforms. After that, the code was written in the same programming language, which received data from the parser and transmitted it to the Automatic Re-activation program, which, with synchronism control, turned on the line after the short circuit disappeared.